api/services/payments/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
python-dotenv==1.0.0
sqlalchemy
asyncpg
alembic
psycopg2-binary
stripe==7.7.0
grpcio 
grpcio-tools
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp